Moved nav icons to navigation folder
