+++ src/components/AdminDash/Navigation/Navigation.js
@@ -9,14 +9,8 @@ import {
   Utilities,
   IconContainer,
   PulseContainer,
-} from './NavigationStyles'
-import {
-  People,
-  Map,
-  DashboardIcon,
-  Pulse,
-  UserShield,
-} from '../AdminDashPics/icons'
+} from './navStyles'
+import { People, Map, DashboardIcon, Pulse, UserShield } from './navIcons'
 import Container from '../UserOptions/OptionContainer'
 
 export default function Navigation() {
