--- src/index.css
@@ -1,31 +1,0 @@
-@import url('https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap');
-
-* {
-  box-sizing: border-box;
-  font-size: 62.5%;
-  /* border: 1px solid red; */
-}
-
-body {
-  font-family: 'Roboto', sans-serif;
-}
-
-span,
-h1,
-h2,
-h3,
-h4,
-h5,
-h6,
-p,
-a,
-address,
-ol,
-ul,
-li,
-form,
-label,
-table,
-caption {
-  font-size: 1.6rem;
-}
