+++ src/index.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import './index.css'
 import App from './App'
 import store from './redux/store'
 import * as Sentry from '@sentry/browser'
